LayoutTests/imported/w3c/web-platform-tests/scroll-animations/view-timelines/block-view-timeline-nested-subject.tentative-expected.txt


FAIL View timeline with subject that is not a direct descendant of the scroll container assert_approx_equals: values do not match for "Timeline's current time at start offset" expected 0 +/- 0.125 but got 150
